requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
cinemagoer==2023.5.1
Django==4.2.1
django-sass==1.1.0
django-sass-processor==1.2.2
greenlet==2.0.2
idna==3.4
IMDbPY==2022.7.9
libsass==0.22.0
lxml==4.9.2
Pillow==9.5.0
psycopg2-binary==2.9.6
requests==2.31.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
tmdbv3api==1.7.7
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
